+++ server/src/main/java/io/github/ralfspoeth/xldr/server/Watcher.java
@@ -2,6 +2,7 @@ package io.github.ralfspoeth.xldr.server;
 
 import io.github.ralfspoeth.filews.DirectoryWatchService;
 import io.github.ralfspoeth.filews.PathEvent;
+import io.github.ralfspoeth.xldr.ldr.Statistics;
 import org.jspecify.annotations.Nullable;
 
 import java.io.Closeable;